requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.45.0
python-dotenv==1.0.0
psycopg2-binary
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1